Bug 1674777 part 1 - Change JS::NewArrayBuffer and JS_New*Array size argument from uint32_t to size_t. r=sfink
A later patch will add a jsapi-test for this.

Differential Revision: https://phabricator.services.mozilla.com/D103273

UltraBlame original commit: d55cbd2be32d19ad66a94ebbd50dab215777e49c

+++ js/public/ArrayBuffer.h
@@ -26,7 +26,7 @@ namespace JS {
 
 
 
-extern JS_PUBLIC_API JSObject* NewArrayBuffer(JSContext* cx, uint32_t nbytes);
+extern JS_PUBLIC_API JSObject* NewArrayBuffer(JSContext* cx, size_t nbytes);
